requirements.txt
flask
supabase>=2
python-dotenv
